reusable useRelationshipHook
